Add files via upload
This project aims to develop a medical diagnosis tool using machine learning techniques to assist in predicting the likelihood of a patient having a specific disease based on their healthcare data. The tool utilizes a dataset containing various health metrics such as glucose levels, blood pressure, BMI, etc., and employs a RandomForestClassifier model to make predictions. The ultimate goal is to provide healthcare professionals with a supplementary tool that can aid in early diagnosis and treatment planning.